+++ packages/playground/src/webapp/app/LoadComponent.tsx
