+++ packages/playground/src/webapp/styles/react-tabs.scss
