+++ packages/playground/src/webapp/styles/toolbarstyle.scss
