+++ packages/playground/src/webapp/toolbars/MyToolbarItemComponent.tsx
